feat(System): Create drawTextureRect default graphic system
I created a drawTextureRect system to draw rectangle of a texture on the screen

+++ Engine/Shared/Component/TextureRectComponent.cpp
@@ -7,5 +7,5 @@
 
 #include "TextureRectComponent.hpp"
 
-TextureRectComponent::TextureRectComponent(std::string path, int left, int top, int width, int height) :
+TextureRectComponent::TextureRectComponent(std::string path, float left, float top, float width, float height) :
     AComponent("TextureRectComponent"), path(path), left(left), top(top), width(width), height(height) {}
